+++ CMakeLists.txt
@@ -31,166 +31,14 @@ endif()
 
 # ---- Declare libraries ----
 
-# -------- graphics ---------
-add_library(
-  bulin_graphics STATIC
-    source/flat_shader.cpp
-    source/shader_model.cpp
-    source/texture.cpp
-)
+add_subdirectory(libs/graphics)
+add_subdirectory(libs/model)
+add_subdirectory(libs/application)
+#add_subdirectory(libs/ui)
 
-set(BULIN_GRAPHICS_EXPORT_FILE_NAME ${CMAKE_CURRENT_BINARY_DIR}/bulin/graphics/export.hpp)
-include(GenerateExportHeader)
-generate_export_header(bulin_graphics EXPORT_FILE_NAME ${BULIN_GRAPHICS_EXPORT_FILE_NAME})
+# ---- Declare executables ----
 
-target_sources(bulin_graphics
-  PUBLIC
-  FILE_SET HEADERS
-  BASE_DIRS
-    include
-    ${CMAKE_CURRENT_BINARY_DIR}
-  FILES
-    include/bulin/graphics/flat_shader.hpp
-    include/bulin/graphics/shader_data.hpp
-    include/bulin/graphics/shader_model.hpp
-    include/bulin/graphics/texture.hpp
-    include/bulin/graphics/types.hpp
-    ${BULIN_GRAPHICS_EXPORT_FILE_NAME}
-)
-
-add_library(bulin::graphics ALIAS bulin_graphics)
-
-target_compile_features(bulin_graphics PUBLIC cxx_std_20)
-
-# platform dependent gl context
-if(CMAKE_SYSTEM_NAME STREQUAL "Windows")
-  set(BULIN_GLCONTEXT WglContext)
-elseif(CMAKE_SYSTEM_NAME STREQUAL "Linux")
-  set(BULIN_GLCONTEXT GlxContext)
-elseif(CMAKE_SYSTEM_NAME STREQUAL "Darwin")
-  set(BULIN_GLCONTEXT CglContext)
-endif()
-
-find_package(Magnum REQUIRED MeshTools Primitives Shaders ${BULIN_GLCONTEXT})
-
-target_link_system_libraries(bulin_graphics
-  PRIVATE
-  Magnum::Magnum
-  Magnum::MeshTools
-  Magnum::Primitives
-  Magnum::Shaders
-  Magnum::GLContext
-)
-
-target_precompile_headers(bulin_graphics
-  PRIVATE
-    <array>
-    <chrono>
-    <string>
-    <map>
-    <variant>
-)
-
-# ------- application -------
-add_library(
-  bulin_application STATIC
-    source/app.cpp
-    source/model.cpp
-)
-
-set(BULIN_APPLICATION_EXPORT_FILE_NAME ${CMAKE_CURRENT_BINARY_DIR}/bulin/application/export.hpp)
-include(GenerateExportHeader)
-generate_export_header(bulin_application EXPORT_FILE_NAME ${BULIN_APPLICATION_EXPORT_FILE_NAME})
-
-target_sources(bulin_application
-  PUBLIC
-  FILE_SET HEADERS
-  BASE_DIRS
-    include
-    ${CMAKE_CURRENT_BINARY_DIR}
-  FILES
-    include/bulin/application/app.hpp
-    include/bulin/application/cereal.hpp
-    include/bulin/application/model.hpp
-    ${BULIN_APPLICATION_EXPORT_FILE_NAME}
-)
-
-add_library(bulin::application ALIAS bulin_application)
-
-target_compile_features(bulin_application PUBLIC cxx_std_20)
-
-find_package(cereal REQUIRED)
-find_package(Immer REQUIRED)
-find_package(Lager REQUIRED)
-
-target_link_system_libraries(bulin_application
-  PUBLIC
-    lager
-  PRIVATE
-    immer
-    cereal::cereal
-)
-
-target_link_libraries(bulin_application
-  PUBLIC
-    bulin::graphics
-)
-
-target_precompile_headers(bulin_application
-  PRIVATE
-    <filesystem>
-    <string>
-    <variant>
-    <lager/extra/struct.hpp>
-    <lager/effect.hpp>
-    <immer/map.hpp>
-)
-
-# ---- Declare executable ----
-find_package(OpenGL REQUIRED)
-find_package(imgui REQUIRED)
-find_package(SDL2 REQUIRED)
-find_path(PORTABLE_FILE_DIALOGS_INCLUDE_DIRS "portable-file-dialogs.h")
-
-add_executable(bulin_exe
-  source/main.cpp
-)
-
-add_executable(bulin::exe ALIAS bulin_exe)
-
-set_property(TARGET bulin_exe PROPERTY OUTPUT_NAME bulin)
-
-target_compile_features(bulin_exe PRIVATE cxx_std_20)
-
-if(WIN32)
-  target_compile_definitions(bulin_exe PRIVATE SDL_MAIN_HANDLED)
-endif()
-
-target_link_system_libraries(bulin_exe
-  PRIVATE
-    OpenGL::GL
-    imgui::imgui
-    $<TARGET_NAME_IF_EXISTS:SDL2::SDL2main>
-    $<IF:$<TARGET_EXISTS:SDL2::SDL2>,SDL2::SDL2,SDL2::SDL2-static>
-    bulin::graphics
-    bulin::application
-)
-
-target_link_libraries(bulin_exe
-  PUBLIC
-    bulin::graphics
-    bulin::application
-)
-
-target_include_directories(bulin_exe PRIVATE ${PORTABLE_FILE_DIALOGS_INCLUDE_DIRS})
-
-target_precompile_headers(bulin_exe
-  PRIVATE
-    <ranges>
-    <iostream>
-    <lager/store.hpp>
-    <lager/event_loop/sdl.hpp>
-)
+add_subdirectory(executables/imgui)
 
 # ---- Install rules ----
 if(NOT CMAKE_SKIP_INSTALL_RULES)
