+++ executables/imgui/main.cpp
@@ -1,6 +1,6 @@
 #include <Magnum/GL/OpenGL.h>
 #include <bulin/application/app.hpp>
-#include <bulin/application/model.hpp>
+#include <bulin/model/model.hpp>
 
 #include <bulin/graphics/shader_data.hpp>
 #include <bulin/graphics/shader_model.hpp>
